Курсор вынесен в отдельный компонент Курсор добавлен в каждый route для стабильной работы
