+++ client/src/components/Cursor/cursor.scss
