+++ client/src/style/style.scss
@@ -5,4 +5,3 @@
 @import './_constants.scss';
 @import './btns.scss';
 @import './form.scss';
-@import './_cursor.scss';
